Naresh update to include grant variations in primary tag
I've updated the formula field and it is now showing the correct value. We need this IF(npsp__Opportunity__r.Account.RecordType.DeveloperName = "Grant_Making_Body","Grants" to show the output as Grants, if the Account Record Type is Grant_Making_Body and the Constituent code is either blank or not trust and Institution.